Fix seed and rendering bug
This update also saves the seed. We generate a short code by converting the current time and use that as a seed which is saved as state so we can save the seed in localstorage for record keeping.

Despite the puzzle being generated, I noticed a different data set being passed to the Board component somehow. I couldn't find where or how that was happening. Either way, we should rely on useEffect here to react when data changes so re-render happens as expected. The key param is now no longer needed on the Board component. By using useEffect on App, we ensure setting the data only when the component is first mounted. This proves to be reliable as I'm not able to replicate anymore that weird seed/rendering bug. The puzzle generated in App is the same data set the Board component now receives.