Fix vertical centering for all screen sizes
If the content exceeds the scroll height, the area wasn't scrolling evenly. The top part of the game grid (cards) was cut off. We were using flex originally but requires a lot more styles to set on various elements. We just need a parent and child to set the appropriate styles to enable vertically centering regardless and if no space to vertically center then content is scrollable with nothing cut off at the top. To do this though with `display: grid;` requires that the parent only has a single child which we have structured with the app root and container.