Simplify styles for vertical scrolling content for modal
We don't need to use flex here as long as we set a fixed height. I don't want the modal going a full 100% height anyway as I want it to appear as a slide in on top from the bottom of the existing content. We just need to set 100% height on element that should be come vertically scrollable if necessary to show all content, otherwise no scrollbars.