Don't read localStorage data on modal open
Before, we were loading the records data from localStorage each time the modal was open. This removes that from the child components so it has no more state and we push that state further up. This is helpful as we only need to load the localStorage once when the app loads. From there on saving each completed game, we update the records state and pass that data as a simple prop down the component tree.