Reset puzzle and wait before generating new one
Due to the CSS transition of flipping the cards, what happened when completing a game is a new puzzle was generated and the cards would show the new value just before flipping over.

Further we don't want the move count changing until its complete so we move the move count reset to the main completion handler.

By delaying enough, the cards will flip with their existing values and then a new puzzle is generated so no flashing the next puzzle values.