Support keyboard input for playing game (partial)
This resolves the Svelte accessibility warnings and adds support for using the keyboard to tab between cards and the space key to flip them, just as you would with mouse input. The first card had a tabindex of 0 but is now 1 because if its 0 it seems to conflict in Safari by jumping to the address bar. So the tab index for the cards is 1 based.